+++ app/src/main/java/woowacourse/omok/domain/exception/RuleViolationException.kt
@@ -1,0 +1,7 @@
+package woowacourse.omok.domain.exception
+
+import woowacourse.omok.domain.rule.lib.type.Violation
+
+class RuleViolationException(
+    violation: Violation,
+) : IllegalArgumentException(violation.message)
