+++ app/src/main/java/woowacourse/omok/domain/rule/lib/other/Iterator.kt
@@ -1,0 +1,7 @@
+package woowacourse.omok.domain.rule.lib.other
+
+interface Iterator<T> {
+    fun hasNext(): Boolean
+
+    fun next(): T
+}
