+++ app/src/main/java/woowacourse/omok/domain/rule/lib/type/Foul.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.domain.rule.lib.type
+
+enum class Foul(
+    val size: Int,
+) {
+    DOUBLE_THREE(3),
+    DOUBLE_FOUR(4),
+}
