+++ app/src/main/java/woowacourse/omok/domain/stone/StoneColor.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.domain.stone
+
+enum class StoneColor {
+    WHITE,
+    BLACK,
+}
