+++ app/src/test/java/woowacourse/omok/StoneFixture.kt
@@ -1,0 +1,489 @@
+package woowacourse.omok
+
+import woowacourse.omok.domain.position.Col
+import woowacourse.omok.domain.position.Position
+import woowacourse.omok.domain.position.Row
+import woowacourse.omok.domain.stone.Stone
+import woowacourse.omok.domain.stone.StoneColor
+
+val STONE_1A_BLACK = Stone(Position(Row.from(1), Col.from('A')), StoneColor.BLACK)
+val STONE_2A_BLACK = Stone(Position(Row.from(2), Col.from('A')), StoneColor.BLACK)
+val STONE_3A_BLACK = Stone(Position(Row.from(3), Col.from('A')), StoneColor.BLACK)
+val STONE_4A_BLACK = Stone(Position(Row.from(4), Col.from('A')), StoneColor.BLACK)
+val STONE_5A_BLACK = Stone(Position(Row.from(5), Col.from('A')), StoneColor.BLACK)
+val STONE_6A_BLACK = Stone(Position(Row.from(6), Col.from('A')), StoneColor.BLACK)
+val STONE_7A_BLACK = Stone(Position(Row.from(7), Col.from('A')), StoneColor.BLACK)
+val STONE_8A_BLACK = Stone(Position(Row.from(8), Col.from('A')), StoneColor.BLACK)
+val STONE_9A_BLACK = Stone(Position(Row.from(9), Col.from('A')), StoneColor.BLACK)
+val STONE_10A_BLACK = Stone(Position(Row.from(10), Col.from('A')), StoneColor.BLACK)
+val STONE_11A_BLACK = Stone(Position(Row.from(11), Col.from('A')), StoneColor.BLACK)
+val STONE_12A_BLACK = Stone(Position(Row.from(12), Col.from('A')), StoneColor.BLACK)
+val STONE_13A_BLACK = Stone(Position(Row.from(13), Col.from('A')), StoneColor.BLACK)
+val STONE_14A_BLACK = Stone(Position(Row.from(14), Col.from('A')), StoneColor.BLACK)
+val STONE_15A_BLACK = Stone(Position(Row.from(15), Col.from('A')), StoneColor.BLACK)
+
+val STONE_1B_BLACK = Stone(Position(Row.from(1), Col.from('B')), StoneColor.BLACK)
+val STONE_2B_BLACK = Stone(Position(Row.from(2), Col.from('B')), StoneColor.BLACK)
+val STONE_3B_BLACK = Stone(Position(Row.from(3), Col.from('B')), StoneColor.BLACK)
+val STONE_4B_BLACK = Stone(Position(Row.from(4), Col.from('B')), StoneColor.BLACK)
+val STONE_5B_BLACK = Stone(Position(Row.from(5), Col.from('B')), StoneColor.BLACK)
+val STONE_6B_BLACK = Stone(Position(Row.from(6), Col.from('B')), StoneColor.BLACK)
+val STONE_7B_BLACK = Stone(Position(Row.from(7), Col.from('B')), StoneColor.BLACK)
+val STONE_8B_BLACK = Stone(Position(Row.from(8), Col.from('B')), StoneColor.BLACK)
+val STONE_9B_BLACK = Stone(Position(Row.from(9), Col.from('B')), StoneColor.BLACK)
+val STONE_10B_BLACK = Stone(Position(Row.from(10), Col.from('B')), StoneColor.BLACK)
+val STONE_11B_BLACK = Stone(Position(Row.from(11), Col.from('B')), StoneColor.BLACK)
+val STONE_12B_BLACK = Stone(Position(Row.from(12), Col.from('B')), StoneColor.BLACK)
+val STONE_13B_BLACK = Stone(Position(Row.from(13), Col.from('B')), StoneColor.BLACK)
+val STONE_14B_BLACK = Stone(Position(Row.from(14), Col.from('B')), StoneColor.BLACK)
+val STONE_15B_BLACK = Stone(Position(Row.from(15), Col.from('B')), StoneColor.BLACK)
+
+val STONE_1C_BLACK = Stone(Position(Row.from(1), Col.from('C')), StoneColor.BLACK)
+val STONE_2C_BLACK = Stone(Position(Row.from(2), Col.from('C')), StoneColor.BLACK)
+val STONE_3C_BLACK = Stone(Position(Row.from(3), Col.from('C')), StoneColor.BLACK)
+val STONE_4C_BLACK = Stone(Position(Row.from(4), Col.from('C')), StoneColor.BLACK)
+val STONE_5C_BLACK = Stone(Position(Row.from(5), Col.from('C')), StoneColor.BLACK)
+val STONE_6C_BLACK = Stone(Position(Row.from(6), Col.from('C')), StoneColor.BLACK)
+val STONE_7C_BLACK = Stone(Position(Row.from(7), Col.from('C')), StoneColor.BLACK)
+val STONE_8C_BLACK = Stone(Position(Row.from(8), Col.from('C')), StoneColor.BLACK)
+val STONE_9C_BLACK = Stone(Position(Row.from(9), Col.from('C')), StoneColor.BLACK)
+val STONE_10C_BLACK = Stone(Position(Row.from(10), Col.from('C')), StoneColor.BLACK)
+val STONE_11C_BLACK = Stone(Position(Row.from(11), Col.from('C')), StoneColor.BLACK)
+val STONE_12C_BLACK = Stone(Position(Row.from(12), Col.from('C')), StoneColor.BLACK)
+val STONE_13C_BLACK = Stone(Position(Row.from(13), Col.from('C')), StoneColor.BLACK)
+val STONE_14C_BLACK = Stone(Position(Row.from(14), Col.from('C')), StoneColor.BLACK)
+val STONE_15C_BLACK = Stone(Position(Row.from(15), Col.from('C')), StoneColor.BLACK)
+
+val STONE_1D_BLACK = Stone(Position(Row.from(1), Col.from('D')), StoneColor.BLACK)
+val STONE_2D_BLACK = Stone(Position(Row.from(2), Col.from('D')), StoneColor.BLACK)
+val STONE_3D_BLACK = Stone(Position(Row.from(3), Col.from('D')), StoneColor.BLACK)
+val STONE_4D_BLACK = Stone(Position(Row.from(4), Col.from('D')), StoneColor.BLACK)
+val STONE_5D_BLACK = Stone(Position(Row.from(5), Col.from('D')), StoneColor.BLACK)
+val STONE_6D_BLACK = Stone(Position(Row.from(6), Col.from('D')), StoneColor.BLACK)
+val STONE_7D_BLACK = Stone(Position(Row.from(7), Col.from('D')), StoneColor.BLACK)
+val STONE_8D_BLACK = Stone(Position(Row.from(8), Col.from('D')), StoneColor.BLACK)
+val STONE_9D_BLACK = Stone(Position(Row.from(9), Col.from('D')), StoneColor.BLACK)
+val STONE_10D_BLACK = Stone(Position(Row.from(10), Col.from('D')), StoneColor.BLACK)
+val STONE_11D_BLACK = Stone(Position(Row.from(11), Col.from('D')), StoneColor.BLACK)
+val STONE_12D_BLACK = Stone(Position(Row.from(12), Col.from('D')), StoneColor.BLACK)
+val STONE_13D_BLACK = Stone(Position(Row.from(13), Col.from('D')), StoneColor.BLACK)
+val STONE_14D_BLACK = Stone(Position(Row.from(14), Col.from('D')), StoneColor.BLACK)
+val STONE_15D_BLACK = Stone(Position(Row.from(15), Col.from('D')), StoneColor.BLACK)
+
+val STONE_1E_BLACK = Stone(Position(Row.from(1), Col.from('E')), StoneColor.BLACK)
+val STONE_2E_BLACK = Stone(Position(Row.from(2), Col.from('E')), StoneColor.BLACK)
+val STONE_3E_BLACK = Stone(Position(Row.from(3), Col.from('E')), StoneColor.BLACK)
+val STONE_4E_BLACK = Stone(Position(Row.from(4), Col.from('E')), StoneColor.BLACK)
+val STONE_5E_BLACK = Stone(Position(Row.from(5), Col.from('E')), StoneColor.BLACK)
+val STONE_6E_BLACK = Stone(Position(Row.from(6), Col.from('E')), StoneColor.BLACK)
+val STONE_7E_BLACK = Stone(Position(Row.from(7), Col.from('E')), StoneColor.BLACK)
+val STONE_8E_BLACK = Stone(Position(Row.from(8), Col.from('E')), StoneColor.BLACK)
+val STONE_9E_BLACK = Stone(Position(Row.from(9), Col.from('E')), StoneColor.BLACK)
+val STONE_10E_BLACK = Stone(Position(Row.from(10), Col.from('E')), StoneColor.BLACK)
+val STONE_11E_BLACK = Stone(Position(Row.from(11), Col.from('E')), StoneColor.BLACK)
+val STONE_12E_BLACK = Stone(Position(Row.from(12), Col.from('E')), StoneColor.BLACK)
+val STONE_13E_BLACK = Stone(Position(Row.from(13), Col.from('E')), StoneColor.BLACK)
+val STONE_14E_BLACK = Stone(Position(Row.from(14), Col.from('E')), StoneColor.BLACK)
+val STONE_15E_BLACK = Stone(Position(Row.from(15), Col.from('E')), StoneColor.BLACK)
+
+val STONE_1F_BLACK = Stone(Position(Row.from(1), Col.from('F')), StoneColor.BLACK)
+val STONE_2F_BLACK = Stone(Position(Row.from(2), Col.from('F')), StoneColor.BLACK)
+val STONE_3F_BLACK = Stone(Position(Row.from(3), Col.from('F')), StoneColor.BLACK)
+val STONE_4F_BLACK = Stone(Position(Row.from(4), Col.from('F')), StoneColor.BLACK)
+val STONE_5F_BLACK = Stone(Position(Row.from(5), Col.from('F')), StoneColor.BLACK)
+val STONE_6F_BLACK = Stone(Position(Row.from(6), Col.from('F')), StoneColor.BLACK)
+val STONE_7F_BLACK = Stone(Position(Row.from(7), Col.from('F')), StoneColor.BLACK)
+val STONE_8F_BLACK = Stone(Position(Row.from(8), Col.from('F')), StoneColor.BLACK)
+val STONE_9F_BLACK = Stone(Position(Row.from(9), Col.from('F')), StoneColor.BLACK)
+val STONE_10F_BLACK = Stone(Position(Row.from(10), Col.from('F')), StoneColor.BLACK)
+val STONE_11F_BLACK = Stone(Position(Row.from(11), Col.from('F')), StoneColor.BLACK)
+val STONE_12F_BLACK = Stone(Position(Row.from(12), Col.from('F')), StoneColor.BLACK)
+val STONE_13F_BLACK = Stone(Position(Row.from(13), Col.from('F')), StoneColor.BLACK)
+val STONE_14F_BLACK = Stone(Position(Row.from(14), Col.from('F')), StoneColor.BLACK)
+val STONE_15F_BLACK = Stone(Position(Row.from(15), Col.from('F')), StoneColor.BLACK)
+
+val STONE_1G_BLACK = Stone(Position(Row.from(1), Col.from('G')), StoneColor.BLACK)
+val STONE_2G_BLACK = Stone(Position(Row.from(2), Col.from('G')), StoneColor.BLACK)
+val STONE_3G_BLACK = Stone(Position(Row.from(3), Col.from('G')), StoneColor.BLACK)
+val STONE_4G_BLACK = Stone(Position(Row.from(4), Col.from('G')), StoneColor.BLACK)
+val STONE_5G_BLACK = Stone(Position(Row.from(5), Col.from('G')), StoneColor.BLACK)
+val STONE_6G_BLACK = Stone(Position(Row.from(6), Col.from('G')), StoneColor.BLACK)
+val STONE_7G_BLACK = Stone(Position(Row.from(7), Col.from('G')), StoneColor.BLACK)
+val STONE_8G_BLACK = Stone(Position(Row.from(8), Col.from('G')), StoneColor.BLACK)
+val STONE_9G_BLACK = Stone(Position(Row.from(9), Col.from('G')), StoneColor.BLACK)
+val STONE_10G_BLACK = Stone(Position(Row.from(10), Col.from('G')), StoneColor.BLACK)
+val STONE_11G_BLACK = Stone(Position(Row.from(11), Col.from('G')), StoneColor.BLACK)
+val STONE_12G_BLACK = Stone(Position(Row.from(12), Col.from('G')), StoneColor.BLACK)
+val STONE_13G_BLACK = Stone(Position(Row.from(13), Col.from('G')), StoneColor.BLACK)
+val STONE_14G_BLACK = Stone(Position(Row.from(14), Col.from('G')), StoneColor.BLACK)
+val STONE_15G_BLACK = Stone(Position(Row.from(15), Col.from('G')), StoneColor.BLACK)
+
+val STONE_1H_BLACK = Stone(Position(Row.from(1), Col.from('H')), StoneColor.BLACK)
+val STONE_2H_BLACK = Stone(Position(Row.from(2), Col.from('H')), StoneColor.BLACK)
+val STONE_3H_BLACK = Stone(Position(Row.from(3), Col.from('H')), StoneColor.BLACK)
+val STONE_4H_BLACK = Stone(Position(Row.from(4), Col.from('H')), StoneColor.BLACK)
+val STONE_5H_BLACK = Stone(Position(Row.from(5), Col.from('H')), StoneColor.BLACK)
+val STONE_6H_BLACK = Stone(Position(Row.from(6), Col.from('H')), StoneColor.BLACK)
+val STONE_7H_BLACK = Stone(Position(Row.from(7), Col.from('H')), StoneColor.BLACK)
+val STONE_8H_BLACK = Stone(Position(Row.from(8), Col.from('H')), StoneColor.BLACK)
+val STONE_9H_BLACK = Stone(Position(Row.from(9), Col.from('H')), StoneColor.BLACK)
+val STONE_10H_BLACK = Stone(Position(Row.from(10), Col.from('H')), StoneColor.BLACK)
+val STONE_11H_BLACK = Stone(Position(Row.from(11), Col.from('H')), StoneColor.BLACK)
+val STONE_12H_BLACK = Stone(Position(Row.from(12), Col.from('H')), StoneColor.BLACK)
+val STONE_13H_BLACK = Stone(Position(Row.from(13), Col.from('H')), StoneColor.BLACK)
+val STONE_14H_BLACK = Stone(Position(Row.from(14), Col.from('H')), StoneColor.BLACK)
+val STONE_15H_BLACK = Stone(Position(Row.from(15), Col.from('H')), StoneColor.BLACK)
+
+val STONE_1I_BLACK = Stone(Position(Row.from(1), Col.from('I')), StoneColor.BLACK)
+val STONE_2I_BLACK = Stone(Position(Row.from(2), Col.from('I')), StoneColor.BLACK)
+val STONE_3I_BLACK = Stone(Position(Row.from(3), Col.from('I')), StoneColor.BLACK)
+val STONE_4I_BLACK = Stone(Position(Row.from(4), Col.from('I')), StoneColor.BLACK)
+val STONE_5I_BLACK = Stone(Position(Row.from(5), Col.from('I')), StoneColor.BLACK)
+val STONE_6I_BLACK = Stone(Position(Row.from(6), Col.from('I')), StoneColor.BLACK)
+val STONE_7I_BLACK = Stone(Position(Row.from(7), Col.from('I')), StoneColor.BLACK)
+val STONE_8I_BLACK = Stone(Position(Row.from(8), Col.from('I')), StoneColor.BLACK)
+val STONE_9I_BLACK = Stone(Position(Row.from(9), Col.from('I')), StoneColor.BLACK)
+val STONE_10I_BLACK = Stone(Position(Row.from(10), Col.from('I')), StoneColor.BLACK)
+val STONE_11I_BLACK = Stone(Position(Row.from(11), Col.from('I')), StoneColor.BLACK)
+val STONE_12I_BLACK = Stone(Position(Row.from(12), Col.from('I')), StoneColor.BLACK)
+val STONE_13I_BLACK = Stone(Position(Row.from(13), Col.from('I')), StoneColor.BLACK)
+val STONE_14I_BLACK = Stone(Position(Row.from(14), Col.from('I')), StoneColor.BLACK)
+val STONE_15I_BLACK = Stone(Position(Row.from(15), Col.from('I')), StoneColor.BLACK)
+
+val STONE_1J_BLACK = Stone(Position(Row.from(1), Col.from('J')), StoneColor.BLACK)
+val STONE_2J_BLACK = Stone(Position(Row.from(2), Col.from('J')), StoneColor.BLACK)
+val STONE_3J_BLACK = Stone(Position(Row.from(3), Col.from('J')), StoneColor.BLACK)
+val STONE_4J_BLACK = Stone(Position(Row.from(4), Col.from('J')), StoneColor.BLACK)
+val STONE_5J_BLACK = Stone(Position(Row.from(5), Col.from('J')), StoneColor.BLACK)
+val STONE_6J_BLACK = Stone(Position(Row.from(6), Col.from('J')), StoneColor.BLACK)
+val STONE_7J_BLACK = Stone(Position(Row.from(7), Col.from('J')), StoneColor.BLACK)
+val STONE_8J_BLACK = Stone(Position(Row.from(8), Col.from('J')), StoneColor.BLACK)
+val STONE_9J_BLACK = Stone(Position(Row.from(9), Col.from('J')), StoneColor.BLACK)
+val STONE_10J_BLACK = Stone(Position(Row.from(10), Col.from('J')), StoneColor.BLACK)
+val STONE_11J_BLACK = Stone(Position(Row.from(11), Col.from('J')), StoneColor.BLACK)
+val STONE_12J_BLACK = Stone(Position(Row.from(12), Col.from('J')), StoneColor.BLACK)
+val STONE_13J_BLACK = Stone(Position(Row.from(13), Col.from('J')), StoneColor.BLACK)
+val STONE_14J_BLACK = Stone(Position(Row.from(14), Col.from('J')), StoneColor.BLACK)
+val STONE_15J_BLACK = Stone(Position(Row.from(15), Col.from('J')), StoneColor.BLACK)
+
+val STONE_1K_BLACK = Stone(Position(Row.from(1), Col.from('K')), StoneColor.BLACK)
+val STONE_2K_BLACK = Stone(Position(Row.from(2), Col.from('K')), StoneColor.BLACK)
+val STONE_3K_BLACK = Stone(Position(Row.from(3), Col.from('K')), StoneColor.BLACK)
+val STONE_4K_BLACK = Stone(Position(Row.from(4), Col.from('K')), StoneColor.BLACK)
+val STONE_5K_BLACK = Stone(Position(Row.from(5), Col.from('K')), StoneColor.BLACK)
+val STONE_6K_BLACK = Stone(Position(Row.from(6), Col.from('K')), StoneColor.BLACK)
+val STONE_7K_BLACK = Stone(Position(Row.from(7), Col.from('K')), StoneColor.BLACK)
+val STONE_8K_BLACK = Stone(Position(Row.from(8), Col.from('K')), StoneColor.BLACK)
+val STONE_9K_BLACK = Stone(Position(Row.from(9), Col.from('K')), StoneColor.BLACK)
+val STONE_10K_BLACK = Stone(Position(Row.from(10), Col.from('K')), StoneColor.BLACK)
+val STONE_11K_BLACK = Stone(Position(Row.from(11), Col.from('K')), StoneColor.BLACK)
+val STONE_12K_BLACK = Stone(Position(Row.from(12), Col.from('K')), StoneColor.BLACK)
+val STONE_13K_BLACK = Stone(Position(Row.from(13), Col.from('K')), StoneColor.BLACK)
+val STONE_14K_BLACK = Stone(Position(Row.from(14), Col.from('K')), StoneColor.BLACK)
+val STONE_15K_BLACK = Stone(Position(Row.from(15), Col.from('K')), StoneColor.BLACK)
+
+val STONE_1L_BLACK = Stone(Position(Row.from(1), Col.from('L')), StoneColor.BLACK)
+val STONE_2L_BLACK = Stone(Position(Row.from(2), Col.from('L')), StoneColor.BLACK)
+val STONE_3L_BLACK = Stone(Position(Row.from(3), Col.from('L')), StoneColor.BLACK)
+val STONE_4L_BLACK = Stone(Position(Row.from(4), Col.from('L')), StoneColor.BLACK)
+val STONE_5L_BLACK = Stone(Position(Row.from(5), Col.from('L')), StoneColor.BLACK)
+val STONE_6L_BLACK = Stone(Position(Row.from(6), Col.from('L')), StoneColor.BLACK)
+val STONE_7L_BLACK = Stone(Position(Row.from(7), Col.from('L')), StoneColor.BLACK)
+val STONE_8L_BLACK = Stone(Position(Row.from(8), Col.from('L')), StoneColor.BLACK)
+val STONE_9L_BLACK = Stone(Position(Row.from(9), Col.from('L')), StoneColor.BLACK)
+val STONE_10L_BLACK = Stone(Position(Row.from(10), Col.from('L')), StoneColor.BLACK)
+val STONE_11L_BLACK = Stone(Position(Row.from(11), Col.from('L')), StoneColor.BLACK)
+val STONE_12L_BLACK = Stone(Position(Row.from(12), Col.from('L')), StoneColor.BLACK)
+val STONE_13L_BLACK = Stone(Position(Row.from(13), Col.from('L')), StoneColor.BLACK)
+val STONE_14L_BLACK = Stone(Position(Row.from(14), Col.from('L')), StoneColor.BLACK)
+val STONE_15L_BLACK = Stone(Position(Row.from(15), Col.from('L')), StoneColor.BLACK)
+
+val STONE_1M_BLACK = Stone(Position(Row.from(1), Col.from('M')), StoneColor.BLACK)
+val STONE_2M_BLACK = Stone(Position(Row.from(2), Col.from('M')), StoneColor.BLACK)
+val STONE_3M_BLACK = Stone(Position(Row.from(3), Col.from('M')), StoneColor.BLACK)
+val STONE_4M_BLACK = Stone(Position(Row.from(4), Col.from('M')), StoneColor.BLACK)
+val STONE_5M_BLACK = Stone(Position(Row.from(5), Col.from('M')), StoneColor.BLACK)
+val STONE_6M_BLACK = Stone(Position(Row.from(6), Col.from('M')), StoneColor.BLACK)
+val STONE_7M_BLACK = Stone(Position(Row.from(7), Col.from('M')), StoneColor.BLACK)
+val STONE_8M_BLACK = Stone(Position(Row.from(8), Col.from('M')), StoneColor.BLACK)
+val STONE_9M_BLACK = Stone(Position(Row.from(9), Col.from('M')), StoneColor.BLACK)
+val STONE_10M_BLACK = Stone(Position(Row.from(10), Col.from('M')), StoneColor.BLACK)
+val STONE_11M_BLACK = Stone(Position(Row.from(11), Col.from('M')), StoneColor.BLACK)
+val STONE_12M_BLACK = Stone(Position(Row.from(12), Col.from('M')), StoneColor.BLACK)
+val STONE_13M_BLACK = Stone(Position(Row.from(13), Col.from('M')), StoneColor.BLACK)
+val STONE_14M_BLACK = Stone(Position(Row.from(14), Col.from('M')), StoneColor.BLACK)
+val STONE_15M_BLACK = Stone(Position(Row.from(15), Col.from('M')), StoneColor.BLACK)
+
+val STONE_1N_BLACK = Stone(Position(Row.from(1), Col.from('N')), StoneColor.BLACK)
+val STONE_2N_BLACK = Stone(Position(Row.from(2), Col.from('N')), StoneColor.BLACK)
+val STONE_3N_BLACK = Stone(Position(Row.from(3), Col.from('N')), StoneColor.BLACK)
+val STONE_4N_BLACK = Stone(Position(Row.from(4), Col.from('N')), StoneColor.BLACK)
+val STONE_5N_BLACK = Stone(Position(Row.from(5), Col.from('N')), StoneColor.BLACK)
+val STONE_6N_BLACK = Stone(Position(Row.from(6), Col.from('N')), StoneColor.BLACK)
+val STONE_7N_BLACK = Stone(Position(Row.from(7), Col.from('N')), StoneColor.BLACK)
+val STONE_8N_BLACK = Stone(Position(Row.from(8), Col.from('N')), StoneColor.BLACK)
+val STONE_9N_BLACK = Stone(Position(Row.from(9), Col.from('N')), StoneColor.BLACK)
+val STONE_10N_BLACK = Stone(Position(Row.from(10), Col.from('N')), StoneColor.BLACK)
+val STONE_11N_BLACK = Stone(Position(Row.from(11), Col.from('N')), StoneColor.BLACK)
+val STONE_12N_BLACK = Stone(Position(Row.from(12), Col.from('N')), StoneColor.BLACK)
+val STONE_13N_BLACK = Stone(Position(Row.from(13), Col.from('N')), StoneColor.BLACK)
+val STONE_14N_BLACK = Stone(Position(Row.from(14), Col.from('N')), StoneColor.BLACK)
+val STONE_15N_BLACK = Stone(Position(Row.from(15), Col.from('N')), StoneColor.BLACK)
+
+val STONE_1O_BLACK = Stone(Position(Row.from(1), Col.from('O')), StoneColor.BLACK)
+val STONE_2O_BLACK = Stone(Position(Row.from(2), Col.from('O')), StoneColor.BLACK)
+val STONE_3O_BLACK = Stone(Position(Row.from(3), Col.from('O')), StoneColor.BLACK)
+val STONE_4O_BLACK = Stone(Position(Row.from(4), Col.from('O')), StoneColor.BLACK)
+val STONE_5O_BLACK = Stone(Position(Row.from(5), Col.from('O')), StoneColor.BLACK)
+val STONE_6O_BLACK = Stone(Position(Row.from(6), Col.from('O')), StoneColor.BLACK)
+val STONE_7O_BLACK = Stone(Position(Row.from(7), Col.from('O')), StoneColor.BLACK)
+val STONE_8O_BLACK = Stone(Position(Row.from(8), Col.from('O')), StoneColor.BLACK)
+val STONE_9O_BLACK = Stone(Position(Row.from(9), Col.from('O')), StoneColor.BLACK)
+val STONE_10O_BLACK = Stone(Position(Row.from(10), Col.from('O')), StoneColor.BLACK)
+val STONE_11O_BLACK = Stone(Position(Row.from(11), Col.from('O')), StoneColor.BLACK)
+val STONE_12O_BLACK = Stone(Position(Row.from(12), Col.from('O')), StoneColor.BLACK)
+val STONE_13O_BLACK = Stone(Position(Row.from(13), Col.from('O')), StoneColor.BLACK)
+val STONE_14O_BLACK = Stone(Position(Row.from(14), Col.from('O')), StoneColor.BLACK)
+val STONE_15O_BLACK = Stone(Position(Row.from(15), Col.from('O')), StoneColor.BLACK)
+
+// WHITE
+
+val STONE_1A_WHITE = Stone(Position(Row.from(1), Col.from('A')), StoneColor.WHITE)
+val STONE_2A_WHITE = Stone(Position(Row.from(2), Col.from('A')), StoneColor.WHITE)
+val STONE_3A_WHITE = Stone(Position(Row.from(3), Col.from('A')), StoneColor.WHITE)
+val STONE_4A_WHITE = Stone(Position(Row.from(4), Col.from('A')), StoneColor.WHITE)
+val STONE_5A_WHITE = Stone(Position(Row.from(5), Col.from('A')), StoneColor.WHITE)
+val STONE_6A_WHITE = Stone(Position(Row.from(6), Col.from('A')), StoneColor.WHITE)
+val STONE_7A_WHITE = Stone(Position(Row.from(7), Col.from('A')), StoneColor.WHITE)
+val STONE_8A_WHITE = Stone(Position(Row.from(8), Col.from('A')), StoneColor.WHITE)
+val STONE_9A_WHITE = Stone(Position(Row.from(9), Col.from('A')), StoneColor.WHITE)
+val STONE_10A_WHITE = Stone(Position(Row.from(10), Col.from('A')), StoneColor.WHITE)
+val STONE_11A_WHITE = Stone(Position(Row.from(11), Col.from('A')), StoneColor.WHITE)
+val STONE_12A_WHITE = Stone(Position(Row.from(12), Col.from('A')), StoneColor.WHITE)
+val STONE_13A_WHITE = Stone(Position(Row.from(13), Col.from('A')), StoneColor.WHITE)
+val STONE_14A_WHITE = Stone(Position(Row.from(14), Col.from('A')), StoneColor.WHITE)
+val STONE_15A_WHITE = Stone(Position(Row.from(15), Col.from('A')), StoneColor.WHITE)
+
+val STONE_1B_WHITE = Stone(Position(Row.from(1), Col.from('B')), StoneColor.WHITE)
+val STONE_2B_WHITE = Stone(Position(Row.from(2), Col.from('B')), StoneColor.WHITE)
+val STONE_3B_WHITE = Stone(Position(Row.from(3), Col.from('B')), StoneColor.WHITE)
+val STONE_4B_WHITE = Stone(Position(Row.from(4), Col.from('B')), StoneColor.WHITE)
+val STONE_5B_WHITE = Stone(Position(Row.from(5), Col.from('B')), StoneColor.WHITE)
+val STONE_6B_WHITE = Stone(Position(Row.from(6), Col.from('B')), StoneColor.WHITE)
+val STONE_7B_WHITE = Stone(Position(Row.from(7), Col.from('B')), StoneColor.WHITE)
+val STONE_8B_WHITE = Stone(Position(Row.from(8), Col.from('B')), StoneColor.WHITE)
+val STONE_9B_WHITE = Stone(Position(Row.from(9), Col.from('B')), StoneColor.WHITE)
+val STONE_10B_WHITE = Stone(Position(Row.from(10), Col.from('B')), StoneColor.WHITE)
+val STONE_11B_WHITE = Stone(Position(Row.from(11), Col.from('B')), StoneColor.WHITE)
+val STONE_12B_WHITE = Stone(Position(Row.from(12), Col.from('B')), StoneColor.WHITE)
+val STONE_13B_WHITE = Stone(Position(Row.from(13), Col.from('B')), StoneColor.WHITE)
+val STONE_14B_WHITE = Stone(Position(Row.from(14), Col.from('B')), StoneColor.WHITE)
+val STONE_15B_WHITE = Stone(Position(Row.from(15), Col.from('B')), StoneColor.WHITE)
+
+val STONE_1C_WHITE = Stone(Position(Row.from(1), Col.from('C')), StoneColor.WHITE)
+val STONE_2C_WHITE = Stone(Position(Row.from(2), Col.from('C')), StoneColor.WHITE)
+val STONE_3C_WHITE = Stone(Position(Row.from(3), Col.from('C')), StoneColor.WHITE)
+val STONE_4C_WHITE = Stone(Position(Row.from(4), Col.from('C')), StoneColor.WHITE)
+val STONE_5C_WHITE = Stone(Position(Row.from(5), Col.from('C')), StoneColor.WHITE)
+val STONE_6C_WHITE = Stone(Position(Row.from(6), Col.from('C')), StoneColor.WHITE)
+val STONE_7C_WHITE = Stone(Position(Row.from(7), Col.from('C')), StoneColor.WHITE)
+val STONE_8C_WHITE = Stone(Position(Row.from(8), Col.from('C')), StoneColor.WHITE)
+val STONE_9C_WHITE = Stone(Position(Row.from(9), Col.from('C')), StoneColor.WHITE)
+val STONE_10C_WHITE = Stone(Position(Row.from(10), Col.from('C')), StoneColor.WHITE)
+val STONE_11C_WHITE = Stone(Position(Row.from(11), Col.from('C')), StoneColor.WHITE)
+val STONE_12C_WHITE = Stone(Position(Row.from(12), Col.from('C')), StoneColor.WHITE)
+val STONE_13C_WHITE = Stone(Position(Row.from(13), Col.from('C')), StoneColor.WHITE)
+val STONE_14C_WHITE = Stone(Position(Row.from(14), Col.from('C')), StoneColor.WHITE)
+val STONE_15C_WHITE = Stone(Position(Row.from(15), Col.from('C')), StoneColor.WHITE)
+
+val STONE_1D_WHITE = Stone(Position(Row.from(1), Col.from('D')), StoneColor.WHITE)
+val STONE_2D_WHITE = Stone(Position(Row.from(2), Col.from('D')), StoneColor.WHITE)
+val STONE_3D_WHITE = Stone(Position(Row.from(3), Col.from('D')), StoneColor.WHITE)
+val STONE_4D_WHITE = Stone(Position(Row.from(4), Col.from('D')), StoneColor.WHITE)
+val STONE_5D_WHITE = Stone(Position(Row.from(5), Col.from('D')), StoneColor.WHITE)
+val STONE_6D_WHITE = Stone(Position(Row.from(6), Col.from('D')), StoneColor.WHITE)
+val STONE_7D_WHITE = Stone(Position(Row.from(7), Col.from('D')), StoneColor.WHITE)
+val STONE_8D_WHITE = Stone(Position(Row.from(8), Col.from('D')), StoneColor.WHITE)
+val STONE_9D_WHITE = Stone(Position(Row.from(9), Col.from('D')), StoneColor.WHITE)
+val STONE_10D_WHITE = Stone(Position(Row.from(10), Col.from('D')), StoneColor.WHITE)
+val STONE_11D_WHITE = Stone(Position(Row.from(11), Col.from('D')), StoneColor.WHITE)
+val STONE_12D_WHITE = Stone(Position(Row.from(12), Col.from('D')), StoneColor.WHITE)
+val STONE_13D_WHITE = Stone(Position(Row.from(13), Col.from('D')), StoneColor.WHITE)
+val STONE_14D_WHITE = Stone(Position(Row.from(14), Col.from('D')), StoneColor.WHITE)
+val STONE_15D_WHITE = Stone(Position(Row.from(15), Col.from('D')), StoneColor.WHITE)
+
+val STONE_1E_WHITE = Stone(Position(Row.from(1), Col.from('E')), StoneColor.WHITE)
+val STONE_2E_WHITE = Stone(Position(Row.from(2), Col.from('E')), StoneColor.WHITE)
+val STONE_3E_WHITE = Stone(Position(Row.from(3), Col.from('E')), StoneColor.WHITE)
+val STONE_4E_WHITE = Stone(Position(Row.from(4), Col.from('E')), StoneColor.WHITE)
+val STONE_5E_WHITE = Stone(Position(Row.from(5), Col.from('E')), StoneColor.WHITE)
+val STONE_6E_WHITE = Stone(Position(Row.from(6), Col.from('E')), StoneColor.WHITE)
+val STONE_7E_WHITE = Stone(Position(Row.from(7), Col.from('E')), StoneColor.WHITE)
+val STONE_8E_WHITE = Stone(Position(Row.from(8), Col.from('E')), StoneColor.WHITE)
+val STONE_9E_WHITE = Stone(Position(Row.from(9), Col.from('E')), StoneColor.WHITE)
+val STONE_10E_WHITE = Stone(Position(Row.from(10), Col.from('E')), StoneColor.WHITE)
+val STONE_11E_WHITE = Stone(Position(Row.from(11), Col.from('E')), StoneColor.WHITE)
+val STONE_12E_WHITE = Stone(Position(Row.from(12), Col.from('E')), StoneColor.WHITE)
+val STONE_13E_WHITE = Stone(Position(Row.from(13), Col.from('E')), StoneColor.WHITE)
+val STONE_14E_WHITE = Stone(Position(Row.from(14), Col.from('E')), StoneColor.WHITE)
+val STONE_15E_WHITE = Stone(Position(Row.from(15), Col.from('E')), StoneColor.WHITE)
+
+val STONE_1F_WHITE = Stone(Position(Row.from(1), Col.from('F')), StoneColor.WHITE)
+val STONE_2F_WHITE = Stone(Position(Row.from(2), Col.from('F')), StoneColor.WHITE)
+val STONE_3F_WHITE = Stone(Position(Row.from(3), Col.from('F')), StoneColor.WHITE)
+val STONE_4F_WHITE = Stone(Position(Row.from(4), Col.from('F')), StoneColor.WHITE)
+val STONE_5F_WHITE = Stone(Position(Row.from(5), Col.from('F')), StoneColor.WHITE)
+val STONE_6F_WHITE = Stone(Position(Row.from(6), Col.from('F')), StoneColor.WHITE)
+val STONE_7F_WHITE = Stone(Position(Row.from(7), Col.from('F')), StoneColor.WHITE)
+val STONE_8F_WHITE = Stone(Position(Row.from(8), Col.from('F')), StoneColor.WHITE)
+val STONE_9F_WHITE = Stone(Position(Row.from(9), Col.from('F')), StoneColor.WHITE)
+val STONE_10F_WHITE = Stone(Position(Row.from(10), Col.from('F')), StoneColor.WHITE)
+val STONE_11F_WHITE = Stone(Position(Row.from(11), Col.from('F')), StoneColor.WHITE)
+val STONE_12F_WHITE = Stone(Position(Row.from(12), Col.from('F')), StoneColor.WHITE)
+val STONE_13F_WHITE = Stone(Position(Row.from(13), Col.from('F')), StoneColor.WHITE)
+val STONE_14F_WHITE = Stone(Position(Row.from(14), Col.from('F')), StoneColor.WHITE)
+val STONE_15F_WHITE = Stone(Position(Row.from(15), Col.from('F')), StoneColor.WHITE)
+
+val STONE_1G_WHITE = Stone(Position(Row.from(1), Col.from('G')), StoneColor.WHITE)
+val STONE_2G_WHITE = Stone(Position(Row.from(2), Col.from('G')), StoneColor.WHITE)
+val STONE_3G_WHITE = Stone(Position(Row.from(3), Col.from('G')), StoneColor.WHITE)
+val STONE_4G_WHITE = Stone(Position(Row.from(4), Col.from('G')), StoneColor.WHITE)
+val STONE_5G_WHITE = Stone(Position(Row.from(5), Col.from('G')), StoneColor.WHITE)
+val STONE_6G_WHITE = Stone(Position(Row.from(6), Col.from('G')), StoneColor.WHITE)
+val STONE_7G_WHITE = Stone(Position(Row.from(7), Col.from('G')), StoneColor.WHITE)
+val STONE_8G_WHITE = Stone(Position(Row.from(8), Col.from('G')), StoneColor.WHITE)
+val STONE_9G_WHITE = Stone(Position(Row.from(9), Col.from('G')), StoneColor.WHITE)
+val STONE_10G_WHITE = Stone(Position(Row.from(10), Col.from('G')), StoneColor.WHITE)
+val STONE_11G_WHITE = Stone(Position(Row.from(11), Col.from('G')), StoneColor.WHITE)
+val STONE_12G_WHITE = Stone(Position(Row.from(12), Col.from('G')), StoneColor.WHITE)
+val STONE_13G_WHITE = Stone(Position(Row.from(13), Col.from('G')), StoneColor.WHITE)
+val STONE_14G_WHITE = Stone(Position(Row.from(14), Col.from('G')), StoneColor.WHITE)
+val STONE_15G_WHITE = Stone(Position(Row.from(15), Col.from('G')), StoneColor.WHITE)
+
+val STONE_1H_WHITE = Stone(Position(Row.from(1), Col.from('H')), StoneColor.WHITE)
+val STONE_2H_WHITE = Stone(Position(Row.from(2), Col.from('H')), StoneColor.WHITE)
+val STONE_3H_WHITE = Stone(Position(Row.from(3), Col.from('H')), StoneColor.WHITE)
+val STONE_4H_WHITE = Stone(Position(Row.from(4), Col.from('H')), StoneColor.WHITE)
+val STONE_5H_WHITE = Stone(Position(Row.from(5), Col.from('H')), StoneColor.WHITE)
+val STONE_6H_WHITE = Stone(Position(Row.from(6), Col.from('H')), StoneColor.WHITE)
+val STONE_7H_WHITE = Stone(Position(Row.from(7), Col.from('H')), StoneColor.WHITE)
+val STONE_8H_WHITE = Stone(Position(Row.from(8), Col.from('H')), StoneColor.WHITE)
+val STONE_9H_WHITE = Stone(Position(Row.from(9), Col.from('H')), StoneColor.WHITE)
+val STONE_10H_WHITE = Stone(Position(Row.from(10), Col.from('H')), StoneColor.WHITE)
+val STONE_11H_WHITE = Stone(Position(Row.from(11), Col.from('H')), StoneColor.WHITE)
+val STONE_12H_WHITE = Stone(Position(Row.from(12), Col.from('H')), StoneColor.WHITE)
+val STONE_13H_WHITE = Stone(Position(Row.from(13), Col.from('H')), StoneColor.WHITE)
+val STONE_14H_WHITE = Stone(Position(Row.from(14), Col.from('H')), StoneColor.WHITE)
+val STONE_15H_WHITE = Stone(Position(Row.from(15), Col.from('H')), StoneColor.WHITE)
+
+val STONE_1I_WHITE = Stone(Position(Row.from(1), Col.from('I')), StoneColor.WHITE)
+val STONE_2I_WHITE = Stone(Position(Row.from(2), Col.from('I')), StoneColor.WHITE)
+val STONE_3I_WHITE = Stone(Position(Row.from(3), Col.from('I')), StoneColor.WHITE)
+val STONE_4I_WHITE = Stone(Position(Row.from(4), Col.from('I')), StoneColor.WHITE)
+val STONE_5I_WHITE = Stone(Position(Row.from(5), Col.from('I')), StoneColor.WHITE)
+val STONE_6I_WHITE = Stone(Position(Row.from(6), Col.from('I')), StoneColor.WHITE)
+val STONE_7I_WHITE = Stone(Position(Row.from(7), Col.from('I')), StoneColor.WHITE)
+val STONE_8I_WHITE = Stone(Position(Row.from(8), Col.from('I')), StoneColor.WHITE)
+val STONE_9I_WHITE = Stone(Position(Row.from(9), Col.from('I')), StoneColor.WHITE)
+val STONE_10I_WHITE = Stone(Position(Row.from(10), Col.from('I')), StoneColor.WHITE)
+val STONE_11I_WHITE = Stone(Position(Row.from(11), Col.from('I')), StoneColor.WHITE)
+val STONE_12I_WHITE = Stone(Position(Row.from(12), Col.from('I')), StoneColor.WHITE)
+val STONE_13I_WHITE = Stone(Position(Row.from(13), Col.from('I')), StoneColor.WHITE)
+val STONE_14I_WHITE = Stone(Position(Row.from(14), Col.from('I')), StoneColor.WHITE)
+val STONE_15I_WHITE = Stone(Position(Row.from(15), Col.from('I')), StoneColor.WHITE)
+
+val STONE_1J_WHITE = Stone(Position(Row.from(1), Col.from('J')), StoneColor.WHITE)
+val STONE_2J_WHITE = Stone(Position(Row.from(2), Col.from('J')), StoneColor.WHITE)
+val STONE_3J_WHITE = Stone(Position(Row.from(3), Col.from('J')), StoneColor.WHITE)
+val STONE_4J_WHITE = Stone(Position(Row.from(4), Col.from('J')), StoneColor.WHITE)
+val STONE_5J_WHITE = Stone(Position(Row.from(5), Col.from('J')), StoneColor.WHITE)
+val STONE_6J_WHITE = Stone(Position(Row.from(6), Col.from('J')), StoneColor.WHITE)
+val STONE_7J_WHITE = Stone(Position(Row.from(7), Col.from('J')), StoneColor.WHITE)
+val STONE_8J_WHITE = Stone(Position(Row.from(8), Col.from('J')), StoneColor.WHITE)
+val STONE_9J_WHITE = Stone(Position(Row.from(9), Col.from('J')), StoneColor.WHITE)
+val STONE_10J_WHITE = Stone(Position(Row.from(10), Col.from('J')), StoneColor.WHITE)
+val STONE_11J_WHITE = Stone(Position(Row.from(11), Col.from('J')), StoneColor.WHITE)
+val STONE_12J_WHITE = Stone(Position(Row.from(12), Col.from('J')), StoneColor.WHITE)
+val STONE_13J_WHITE = Stone(Position(Row.from(13), Col.from('J')), StoneColor.WHITE)
+val STONE_14J_WHITE = Stone(Position(Row.from(14), Col.from('J')), StoneColor.WHITE)
+val STONE_15J_WHITE = Stone(Position(Row.from(15), Col.from('J')), StoneColor.WHITE)
+
+val STONE_1K_WHITE = Stone(Position(Row.from(1), Col.from('K')), StoneColor.WHITE)
+val STONE_2K_WHITE = Stone(Position(Row.from(2), Col.from('K')), StoneColor.WHITE)
+val STONE_3K_WHITE = Stone(Position(Row.from(3), Col.from('K')), StoneColor.WHITE)
+val STONE_4K_WHITE = Stone(Position(Row.from(4), Col.from('K')), StoneColor.WHITE)
+val STONE_5K_WHITE = Stone(Position(Row.from(5), Col.from('K')), StoneColor.WHITE)
+val STONE_6K_WHITE = Stone(Position(Row.from(6), Col.from('K')), StoneColor.WHITE)
+val STONE_7K_WHITE = Stone(Position(Row.from(7), Col.from('K')), StoneColor.WHITE)
+val STONE_8K_WHITE = Stone(Position(Row.from(8), Col.from('K')), StoneColor.WHITE)
+val STONE_9K_WHITE = Stone(Position(Row.from(9), Col.from('K')), StoneColor.WHITE)
+val STONE_10K_WHITE = Stone(Position(Row.from(10), Col.from('K')), StoneColor.WHITE)
+val STONE_11K_WHITE = Stone(Position(Row.from(11), Col.from('K')), StoneColor.WHITE)
+val STONE_12K_WHITE = Stone(Position(Row.from(12), Col.from('K')), StoneColor.WHITE)
+val STONE_13K_WHITE = Stone(Position(Row.from(13), Col.from('K')), StoneColor.WHITE)
+val STONE_14K_WHITE = Stone(Position(Row.from(14), Col.from('K')), StoneColor.WHITE)
+val STONE_15K_WHITE = Stone(Position(Row.from(15), Col.from('K')), StoneColor.WHITE)
+
+val STONE_1L_WHITE = Stone(Position(Row.from(1), Col.from('L')), StoneColor.WHITE)
+val STONE_2L_WHITE = Stone(Position(Row.from(2), Col.from('L')), StoneColor.WHITE)
+val STONE_3L_WHITE = Stone(Position(Row.from(3), Col.from('L')), StoneColor.WHITE)
+val STONE_4L_WHITE = Stone(Position(Row.from(4), Col.from('L')), StoneColor.WHITE)
+val STONE_5L_WHITE = Stone(Position(Row.from(5), Col.from('L')), StoneColor.WHITE)
+val STONE_6L_WHITE = Stone(Position(Row.from(6), Col.from('L')), StoneColor.WHITE)
+val STONE_7L_WHITE = Stone(Position(Row.from(7), Col.from('L')), StoneColor.WHITE)
+val STONE_8L_WHITE = Stone(Position(Row.from(8), Col.from('L')), StoneColor.WHITE)
+val STONE_9L_WHITE = Stone(Position(Row.from(9), Col.from('L')), StoneColor.WHITE)
+val STONE_10L_WHITE = Stone(Position(Row.from(10), Col.from('L')), StoneColor.WHITE)
+val STONE_11L_WHITE = Stone(Position(Row.from(11), Col.from('L')), StoneColor.WHITE)
+val STONE_12L_WHITE = Stone(Position(Row.from(12), Col.from('L')), StoneColor.WHITE)
+val STONE_13L_WHITE = Stone(Position(Row.from(13), Col.from('L')), StoneColor.WHITE)
+val STONE_14L_WHITE = Stone(Position(Row.from(14), Col.from('L')), StoneColor.WHITE)
+val STONE_15L_WHITE = Stone(Position(Row.from(15), Col.from('L')), StoneColor.WHITE)
+
+val STONE_1M_WHITE = Stone(Position(Row.from(1), Col.from('M')), StoneColor.WHITE)
+val STONE_2M_WHITE = Stone(Position(Row.from(2), Col.from('M')), StoneColor.WHITE)
+val STONE_3M_WHITE = Stone(Position(Row.from(3), Col.from('M')), StoneColor.WHITE)
+val STONE_4M_WHITE = Stone(Position(Row.from(4), Col.from('M')), StoneColor.WHITE)
+val STONE_5M_WHITE = Stone(Position(Row.from(5), Col.from('M')), StoneColor.WHITE)
+val STONE_6M_WHITE = Stone(Position(Row.from(6), Col.from('M')), StoneColor.WHITE)
+val STONE_7M_WHITE = Stone(Position(Row.from(7), Col.from('M')), StoneColor.WHITE)
+val STONE_8M_WHITE = Stone(Position(Row.from(8), Col.from('M')), StoneColor.WHITE)
+val STONE_9M_WHITE = Stone(Position(Row.from(9), Col.from('M')), StoneColor.WHITE)
+val STONE_10M_WHITE = Stone(Position(Row.from(10), Col.from('M')), StoneColor.WHITE)
+val STONE_11M_WHITE = Stone(Position(Row.from(11), Col.from('M')), StoneColor.WHITE)
+val STONE_12M_WHITE = Stone(Position(Row.from(12), Col.from('M')), StoneColor.WHITE)
+val STONE_13M_WHITE = Stone(Position(Row.from(13), Col.from('M')), StoneColor.WHITE)
+val STONE_14M_WHITE = Stone(Position(Row.from(14), Col.from('M')), StoneColor.WHITE)
+val STONE_15M_WHITE = Stone(Position(Row.from(15), Col.from('M')), StoneColor.WHITE)
+
+val STONE_1N_WHITE = Stone(Position(Row.from(1), Col.from('N')), StoneColor.WHITE)
+val STONE_2N_WHITE = Stone(Position(Row.from(2), Col.from('N')), StoneColor.WHITE)
+val STONE_3N_WHITE = Stone(Position(Row.from(3), Col.from('N')), StoneColor.WHITE)
+val STONE_4N_WHITE = Stone(Position(Row.from(4), Col.from('N')), StoneColor.WHITE)
+val STONE_5N_WHITE = Stone(Position(Row.from(5), Col.from('N')), StoneColor.WHITE)
+val STONE_6N_WHITE = Stone(Position(Row.from(6), Col.from('N')), StoneColor.WHITE)
+val STONE_7N_WHITE = Stone(Position(Row.from(7), Col.from('N')), StoneColor.WHITE)
+val STONE_8N_WHITE = Stone(Position(Row.from(8), Col.from('N')), StoneColor.WHITE)
+val STONE_9N_WHITE = Stone(Position(Row.from(9), Col.from('N')), StoneColor.WHITE)
+val STONE_10N_WHITE = Stone(Position(Row.from(10), Col.from('N')), StoneColor.WHITE)
+val STONE_11N_WHITE = Stone(Position(Row.from(11), Col.from('N')), StoneColor.WHITE)
+val STONE_12N_WHITE = Stone(Position(Row.from(12), Col.from('N')), StoneColor.WHITE)
+val STONE_13N_WHITE = Stone(Position(Row.from(13), Col.from('N')), StoneColor.WHITE)
+val STONE_14N_WHITE = Stone(Position(Row.from(14), Col.from('N')), StoneColor.WHITE)
+val STONE_15N_WHITE = Stone(Position(Row.from(15), Col.from('N')), StoneColor.WHITE)
+
+val STONE_1O_WHITE = Stone(Position(Row.from(1), Col.from('O')), StoneColor.WHITE)
+val STONE_2O_WHITE = Stone(Position(Row.from(2), Col.from('O')), StoneColor.WHITE)
+val STONE_3O_WHITE = Stone(Position(Row.from(3), Col.from('O')), StoneColor.WHITE)
+val STONE_4O_WHITE = Stone(Position(Row.from(4), Col.from('O')), StoneColor.WHITE)
+val STONE_5O_WHITE = Stone(Position(Row.from(5), Col.from('O')), StoneColor.WHITE)
+val STONE_6O_WHITE = Stone(Position(Row.from(6), Col.from('O')), StoneColor.WHITE)
+val STONE_7O_WHITE = Stone(Position(Row.from(7), Col.from('O')), StoneColor.WHITE)
+val STONE_8O_WHITE = Stone(Position(Row.from(8), Col.from('O')), StoneColor.WHITE)
+val STONE_9O_WHITE = Stone(Position(Row.from(9), Col.from('O')), StoneColor.WHITE)
+val STONE_10O_WHITE = Stone(Position(Row.from(10), Col.from('O')), StoneColor.WHITE)
+val STONE_11O_WHITE = Stone(Position(Row.from(11), Col.from('O')), StoneColor.WHITE)
+val STONE_12O_WHITE = Stone(Position(Row.from(12), Col.from('O')), StoneColor.WHITE)
+val STONE_13O_WHITE = Stone(Position(Row.from(13), Col.from('O')), StoneColor.WHITE)
+val STONE_14O_WHITE = Stone(Position(Row.from(14), Col.from('O')), StoneColor.WHITE)
+val STONE_15O_WHITE = Stone(Position(Row.from(15), Col.from('O')), StoneColor.WHITE)
