--- src/main/kotlin/domain/stone/StoneColor.kt
@@ -1,6 +1,0 @@
-package domain.stone
-
-enum class StoneColor {
-    WHITE,
-    BLACK,
-}
