--- src/main/kotlin/exception/RuleViolationException.kt
@@ -1,7 +1,0 @@
-package exception
-
-import rule.lib.type.Violation
-
-class RuleViolationException(
-    violation: Violation,
-) : IllegalArgumentException(violation.message)
