--- src/main/kotlin/rule/lib/other/Iterator.kt
@@ -1,7 +1,0 @@
-package rule.lib.other
-
-interface Iterator<T> {
-    fun hasNext(): Boolean
-
-    fun next(): T
-}
