--- src/main/kotlin/rule/lib/type/Foul.kt
@@ -1,8 +1,0 @@
-package rule.lib.type
-
-enum class Foul(
-    val size: Int,
-) {
-    DOUBLE_THREE(3),
-    DOUBLE_FOUR(4),
-}
