--- src/test/kotlin/ColTest.kt
@@ -1,38 +1,0 @@
-import domain.position.Col
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertFalse
-import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertAll
-import org.junit.jupiter.api.assertDoesNotThrow
-import org.junit.jupiter.api.assertThrows
-
-class ColTest {
-    @Test
-    fun `유효한 열이 아니라면 예외가 발생한다`() {
-        assertThrows<IllegalArgumentException> { Col.from('Z') }
-    }
-
-    @Test
-    fun `열이 유효하면 생성한다`() {
-        assertDoesNotThrow { Col.from('A') }
-    }
-
-    @Test
-    fun `A를 1로 계산한다`() {
-        val col = Col.from('A')
-        assertEquals(col.value, 1)
-    }
-
-    @Test
-    fun `같은 열인지 판단한다`() {
-        val col = Col.from('A')
-        val sameCol = Col.from('A')
-        val otherCol = Col.from('B')
-
-        assertAll(
-            { assertTrue(col.isSame(sameCol)) },
-            { assertFalse(col.isSame(otherCol)) },
-        )
-    }
-}
